test_requirements.txt
fabric3
paramiko<2.0.0
coverage
tox
django-libs
mixer
mock
flake8
ipdb
